requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiohttp-cors==0.7.0
aiohttp-jinja2==1.5
aiosignal==1.2.0
apispec==5.2.2
astroid==2.11.7
async-timeout==4.0.2
attrs==21.4.0
black==22.6.0
Brotli==1.0.9
cchardet==2.1.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
dacite==1.6.0
defaultlist==1.0.0
dill==0.3.5.1
freezegun==1.2.1
frozenlist==1.3.0
idna==3.3
iniconfig==1.1.1
isort==5.10.1
Jinja2==3.1.2
ktstools==0.10.9
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-enum==1.5.1
mccabe==0.7.0
multidict==6.0.2
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
prometheus-client==0.14.1
py==1.11.0
pycares==4.2.1
pycparser==2.21
PyJWT==2.4.0
pylint==2.14.4
pyparsing==3.0.9
pytest==7.1.2
pytest-aiohttp==1.0.4
pytest-asyncio==0.19.0
pytest-env==0.6.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
sentry-sdk==1.7.1
six==1.16.0
swagger-ui-bundle==0.0.9
tomli==2.0.1
tomlkit==0.11.1
typing-inspect==0.7.1
typing_extensions==4.3.0
urllib3==1.26.10
wrapt==1.14.1
yarl==1.7.2
